requirements.txt
discord.py
python-dotenv
requests
diffusers
invisible_watermark 
transformers 
accelerate 
safetensors